refactor: move GlobalConfig to format package
